+++ client/projects/client/src/app/lib/services/xmpp-events-distribution.service.ts
@@ -1,9 +1,14 @@
-import { DestroyRef, Injectable, OnDestroy } from "@angular/core";
-import { AppState } from "../../store/store";
+import { Injectable, OnDestroy } from "@angular/core";
+import { AppState } from "@store/store";
 import { Store } from "@ngrx/store";
-import { filter, map, Observable, of, Subscription, switchMap } from "rxjs";
+import { filter, map, Observable, of, Subscription, switchMap, take, takeLast, tap } from "rxjs";
 import { ReceivedIQ, ReceivedMessage, ReceivedPresence } from "stanza/protocol";
-import { DistributionService } from "@lib";
+import { XmppService } from "./xmpp.service";
+import { IQManager } from "./xmpp/iq";
+import { ChatMessageAdd } from "@store/chat/actions";
+import { Message } from "@lib";
+import { MessageManager } from "./xmpp/message";
+
 
 /**
  *
@@ -16,55 +21,80 @@ import { DistributionService } from "@lib";
  * @class EventsDistributionService
  */
 @Injectable()
-export class XmppEventsDistributionService implements DistributionService, OnDestroy {
-    subscription: Subscription = new Subscription
-    constructor(private store: Store<AppState>, private ref: DestroyRef) { }
+export class XmppEventsDistributionService implements OnDestroy {
+    sub: Subscription = new Subscription
+    constructor(private store: Store<AppState>) { }
 
     ngOnDestroy(): void {
-        this.subscription.unsubscribe()
+        this.sub.unsubscribe()
+    }
+
+    start(xmpp: XmppService) {
+        this.sub.add(
+            xmpp.connected$.pipe(
+                filter(connect => connect),
+                take(1),
+                tap(() => {
+
+                    // run initial queries
+                    this.initialQueries(xmpp);
+
+                    // provide event sources to distribution service
+                    this.distribute(xmpp);
+                }),
+            ).subscribe()
+        )
+    }
+
+    initialQueries(xmpp: XmppService): void {
+        // get static list of rooms
+        this.sub.add(
+            xmpp.queryRoomsOnline().subscribe()
+        );
+
+        // pubsub
+        this.sub.add(
+            xmpp.getPubsub().subscribe()
+        );
+
+        // contact list
+        this.sub.add(
+            xmpp.getRoster().subscribe()
+        );
     }
 
     /**
      *
      * Distribute and store events data from xmpp
      *
-     * @param {Observable<ReceivedMessage>} messages$
-     * @param {Observable<ReceivedPresence>} presence$
-     * @param {Observable<ReceivedIQ>} iq$
-     * @memberof XmppEventsDistributionService
      */
-    run(
-        messages$: Observable<ReceivedMessage>,
-        presence$: Observable<ReceivedPresence>,
-        iq$: Observable<ReceivedIQ>,
-    ): void {
-        // this.subscription.add(
-        //     messages$.pipe(
-        //         switchMap((message: ReceivedMessage) => {
-        //             if (message.type !== MessageType.GroupChat) {
-        //                 return this.store.select(selectConversationByJid(message.from)).pipe(
-        //                     map((conversation) => {
-        //                         if (conversation && conversation.length === 0) {
-        //                             this.store.dispatch(ChatConversationAdd({ payload: { jid: message.from} }))
-        //                         }
-        //                         return message;
-        //                     })
-        //                 );
-        //             }
-        //             return of(message);
-        //         }),
-        //         filter(message => message.body !== undefined)
-        //     ).subscribe((message) => this.store.dispatch(ChatMessageAdd({ payload: message }))),
-        // );
+    distribute(xmpp: XmppService): void {
+        if (!xmpp) return
+        this.sub.add(
+            xmpp.receivedIQ$.pipe(
+                switchMap((iq: ReceivedIQ, _) => {
+                    const ctx = new IQManager(this.store, xmpp, iq);
+                    return ctx.handle();    
+                })
+            ).subscribe()
+        );
 
-        // this.subscription.add(
-        //     presence$.pipe(
-        //         map(presence => {
-        //             if (presence.type && SubscriptionTypes.includes(presence.type)) {
-        //                 this.store.dispatch(ChatSubscriptionAdd({ payload: presence }))
-        //             }
-        //         })
-        //     ).subscribe(),
-        // );
+        this.sub.add(
+            xmpp.receivedMessage$.subscribe(msg => {
+                const ctx = new MessageManager(this.store, xmpp, msg);
+                ctx.handle();
+            })
+        );
+
     }
+
+    // this.subscription.add(
+    //     presence$.pipe(
+    //         map(presence => {
+    //             if (presence.type && SubscriptionTypes.includes(presence.type)) {
+    //                 this.store.dispatch(ChatSubscriptionAdd({ payload: presence }))
+    //             }
+    //         })
+    //     ).subscribe(),
+    // );
 }