+++ client/projects/client/src/app/modules/chat/components/conversation-area/chat-menu/private-room-item.component.ts
@@ -7,21 +7,10 @@ import { Room } from "@lib";
     selector: "chat-private-room-item",
     template: `
 <div class="flex items-center justify-between gap-2">
-    <div class="flex gap-2 p-2 cursor-pointer" (click)="handleClick()">
-        @if (item.image) {
-            <img class="w-6 h-6 rounded-full shrink-0" src="{{ item.image }}" />
-        } @else {
-            <div class="w-6 h-6 bg-slate-300 rounded-full flex items-center justify-center">
-                <i class="fa fa-user text-slate-500"></i>
-            </div>
-        }
+    <div class="flex items-center gap-2 p-2 cursor-pointer" (click)="handleClick()">
+        <ui-avatar [image]="item.image" [size]="8" />
 
         {{ item.jid | subroomname }}
-        <!-- @if (item.name) {
-            {{ item.name }}
-        } @else {
-            {{ item.jid | subroomname }}
-        } -->
     </div>
     <button class="collocutor-menu bg-slate-100 rounded-lg px-2">
         <i class="fa-solid fa-ellipsis"></i>