+++ client/projects/client/src/app/modules/chat/main-chat.component.ts
@@ -1,6 +1,11 @@
 import { Component } from "@angular/core";
+import { Store } from "@ngrx/store";
+import { AppState } from "@store/store";
 import { UiService } from "@ui";
-import { Observable } from "rxjs";
+import { Observable, of, switchMap } from "rxjs";
+import { XmppService } from "../../lib/services/xmpp.service";
+import { XmppEventsDistributionService } from "../../lib/services/xmpp-events-distribution.service";
+import { takeUntilDestroyed } from "@angular/core/rxjs-interop";
 
 @Component({
     selector: "app-chat",
@@ -11,9 +16,14 @@ export class MainChatComponent {
 
     constructor(
         private ui: UiService,
+        private store: Store<AppState>,
+        private xmpp: XmppService,
+        private eventsDistribution: XmppEventsDistributionService,
     ) {
         this.ui.title.next("Chat");
         this.loading$ = this.ui.appLoading.asObservable();
-    }
 
+
+        this.eventsDistribution.start(this.xmpp);
+    }
 }