+++ client/projects/ui/src/lib/ui.module.ts
@@ -1,18 +1,20 @@
 import { NgModule } from '@angular/core';
 import { FormMessagesComponent } from './form-messages/form-messages.component';
 import { KratosInputComponent } from './kratos-input/kratos-input.component';
-
+import { UiAvatarComponent } from './avatar.component';
 
 @NgModule({
+    imports: [],
     declarations: [
         FormMessagesComponent,
         KratosInputComponent,
-    ],
-    imports: [
+        UiAvatarComponent,
     ],
     exports: [
         FormMessagesComponent,
         KratosInputComponent,
+        UiAvatarComponent,
     ]
 })
 export class UiModule { }
+